CLEAN: Removed the election Data Story.
This was using data published at a VU endpoint, created during a VU
outig last year.  The Data Story is not such a good fit for Labs to
begin with, since the second query, written by a data journalist,
could be interpreted politically, and because the first query relies
on Google Chart support, which is currently disabled on Labs.